# chess-backend/docs/api-contract/openapi.yaml
openapi: "3.0.3"
info:
  title: Chess99 API
  description: |
    REST API for the Chess99 real-time multiplayer chess platform.

    Covers authentication, game management, ratings, tournaments,
    tutorials, social features, WebSocket lifecycle endpoints, and
    push-notification device registration.

    All authenticated endpoints require a Bearer token obtained from
    the `/auth/login`, `/auth/register`, or OAuth endpoints.
    Tokens are Laravel Sanctum personal-access tokens.
  version: "1.0.0"
  contact:
    name: Chess99 Engineering
    url: https://chess99.com
  license:
    name: Proprietary
servers:
  - url: https://api.chess99.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Laravel Sanctum personal-access token

  # ── Reusable Parameters ───────────────────────────────────────────────────
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Magnus"
        email:
          type: string
          format: email
          example: "magnus@chess99.com"
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: "https://api.chess99.com/storage/avatars/1_1700000000.jpg"
        rating:
          type: integer
          example: 1500
        is_provisional:
          type: boolean
          example: true
        games_played:
          type: integer
          example: 12
        peak_rating:
          type: integer
          example: 1520
        tutorial_xp:
          type: integer
          example: 350
        tutorial_level:
          type: integer
          example: 3
        current_skill_tier:
          type: string
          enum: [beginner, intermediate, advanced]
          nullable: true
          example: "beginner"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      description: Lightweight user reference used inside game objects
      properties:
        id:
          type: integer
        name:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
        rating:
          type: integer
...
    SuccessMessage:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "The given data was invalid."
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["The email field is required."]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Unauthenticated."
        status:
          type: string
          example: "error"
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: "Unauthenticated."
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: "Resource not found."
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
...
paths:
...
  /profile:
    post:
      operationId: updateProfile
      tags: [User]
      summary: Update user profile
      description: Supports multipart form data for avatar uploads.
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                avatar:
                  type: string
                  format: binary
                  description: Avatar image file (JPEG, PNG, max 2MB)
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── Friends ──────────────────────────────────────────────────────────────
  /friends:
    get:
      operationId: getFriends
      tags: [Friends]
      summary: List accepted friends
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Friend list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /friends/{friendId}:
    post:
      operationId: addFriend
      tags: [Friends]
      summary: Send a friend request
      security:
        - BearerAuth: []
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeFriend
      tags: [Friends]
      summary: Remove a friend
      security:
        - BearerAuth: []
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friend removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
...